Move common functions and unet model from generalization.py to its packages
